Began refactoring for mdtraj code
