+++ backmap/__init__.py
@@ -5,7 +5,6 @@ Backmapping for pi-conjugated peptides
 
 # Add imports here
 from .backmapping import *
-from .topology import *
 
 # Handle versioneer
 from ._version import get_versions
