+++ backmap/data/residue_mappings/ALA.pdb
@@ -1,0 +1,26 @@
+COMPND    UNNAMED
+AUTHOR    GENERATED BY OPEN BABEL 2.3.2
+ATOM      1  C   ALA A   1      -3.842   2.862  -0.266  1.00  0.00           C
+ATOM      2  CA  ALA A   1      -2.537   3.401   0.241  1.00  0.00           C  
+ATOM      3  CB  ALA A   1      -2.464   4.910   0.012  1.00  0.00           C  
+ATOM      4  HB1 ALA A   1      -2.530   5.148  -1.072  1.00  0.00           H  
+ATOM      5  HB2 ALA A   1      -1.504   5.306   0.406  1.00  0.00           H  
+ATOM      6  HB3 ALA A   1      -3.298   5.420   0.542  1.00  0.00           H  
+ATOM      7  N   ALA A   1      -1.422   2.733  -0.423  1.00  0.00           N  
+ATOM      8  HA  ALA A   1      -2.471   3.205   1.333  1.00  0.00           H  
+ATOM      9  H   ALA A   1      -1.479   2.906  -1.454  1.00  0.00           H  
+ATOM     10  O   ALA A   1      -4.112   1.677  -0.150  1.00  0.00           O  
+CONECT    1   10    2                                                 
+CONECT    2    7    1    3    8                                       
+CONECT    2                                                           
+CONECT    3    4    2    5    6                                       
+CONECT    3                                                           
+CONECT    4    3                                                      
+CONECT    5    3                                                      
+CONECT    6    3                                                      
+CONECT    7    9    2                                                 
+CONECT    8    2                                                      
+CONECT    9    7                                                      
+CONECT   10    1                                                      
+MASTER        0    0    0    0    0    0    0    0   10    0   10    0
+END
