+++ backmap/tests/MOL_GMX.gro
@@ -1,0 +1,149 @@
+MOL_GMX.gro created by acpype (Rev: 0) on Wed Aug 21 16:07:53 2019
+ 146
+    1  MOL   O1    1   2.164  -0.414  -0.185
+    1  MOL   C1    2   2.085  -0.355  -0.282
+    1  MOL   C2    3   2.010  -0.234  -0.231
+    1  MOL   C3    4   2.053  -0.170  -0.097
+    1  MOL   C4    5   2.190  -0.099  -0.104
+    1  MOL   O2    6   2.297  -0.178  -0.121
+    1  MOL   O3    7   2.077  -0.401  -0.392
+    1  MOL   O4    8   2.201   0.021  -0.096
+    1  MOL   N1    9   1.957  -0.069  -0.055
+    1  MOL   C5   10   1.836  -0.102  -0.005
+    1  MOL   C6   11   1.752   0.014   0.050
+    1  MOL   O5   12   1.794  -0.218  -0.001
+    1  MOL   N2   13   1.613  -0.024   0.040
+    1  MOL   C7   14   1.795   0.043   0.197
+    1  MOL   C8   15   1.718   0.155   0.263
+    1  MOL   C9   16   1.609   0.128   0.346
+    1  MOL  C10   17   1.754   0.289   0.241
+    1  MOL  C11   18   1.537   0.232   0.405
+    1  MOL  C12   19   1.682   0.393   0.300
+    1  MOL  C13   20   1.573   0.364   0.382
+    1  MOL  C14   21   1.519   0.056  -0.018
+    1  MOL  C15   22   1.381  -0.010  -0.038
+    1  MOL  C16   23   1.385  -0.114  -0.151
+    1  MOL   O6   24   1.541   0.171  -0.054
+    1  MOL   N3   25   1.288   0.098  -0.066
+    1  MOL  C17   26   1.154   0.085  -0.050
+    1  MOL  C18   27   1.073   0.211  -0.078
+    1  MOL   O7   28   1.101  -0.021  -0.015
+    1  MOL   N4   29   0.934   0.182  -0.049
+    1  MOL  C19   30   0.835   0.265  -0.093
+    1  MOL  C20   31   0.694   0.221  -0.067
+    1  MOL   O8   32   0.860   0.371  -0.150
+    1  MOL  C21   33   0.592   0.285  -0.137
+    1  MOL  C22   34   0.459   0.248  -0.119
+    1  MOL  C23   35   0.423   0.147  -0.028
+    1  MOL  C24   36   0.526   0.084   0.044
+    1  MOL  C25   37   0.659   0.120   0.025
+    1  MOL  C26   38   0.281   0.113  -0.013
+    1  MOL  C27   39   0.229   0.016   0.065
+    1  MOL  C28   40   0.087  -0.018   0.080
+    1  MOL  C29   41  -0.017   0.050   0.014
+    1  MOL  C30   42  -0.150   0.013   0.031
+    1  MOL  C31   43  -0.186  -0.094   0.116
+    1  MOL  C32   44  -0.082  -0.161   0.183
+    1  MOL  C33   45   0.051  -0.124   0.165
+    1  MOL  C34   46  -0.324  -0.137   0.139
+    1  MOL  C35   47  -0.435  -0.086   0.083
+    1  MOL  C36   48  -0.573  -0.130   0.105
+    1  MOL  C37   49  -0.677  -0.066   0.035
+    1  MOL  C38   50  -0.810  -0.103   0.052
+    1  MOL  C39   51  -0.844  -0.207   0.139
+    1  MOL  C40   52  -0.741  -0.271   0.211
+    1  MOL  C41   53  -0.609  -0.233   0.194
+    1  MOL  C42   54  -0.984  -0.254   0.163
+    1  MOL   O9   55  -1.014  -0.324   0.260
+    1  MOL   N5   56  -1.078  -0.215   0.072
+    1  MOL  C43   57  -1.213  -0.265   0.081
+    1  MOL  C44   58  -1.285  -0.238  -0.051
+    1  MOL  O10   59  -1.227  -0.189  -0.147
+    1  MOL   N6   60  -1.418  -0.271  -0.052
+    1  MOL  C45   61  -1.498  -0.271  -0.174
+    1  MOL  C46   62  -1.532  -0.413  -0.220
+    1  MOL  C47   63  -1.627  -0.187  -0.161
+    1  MOL  O11   64  -1.719  -0.200  -0.241
+    1  MOL   N7   65  -1.629  -0.098  -0.057
+    1  MOL  C48   66  -1.738  -0.004  -0.039
+    1  MOL  C49   67  -1.678   0.137  -0.029
+    1  MOL  O12   68  -1.569   0.156   0.025
+    1  MOL  C50   69  -1.819  -0.033   0.091
+    1  MOL  C51   70  -1.873  -0.174   0.098
+    1  MOL  C52   71  -1.983  -0.213   0.021
+    1  MOL  C53   72  -1.813  -0.269   0.182
+    1  MOL  C54   73  -1.861  -0.400   0.188
+    1  MOL  C55   74  -2.031  -0.344   0.026
+    1  MOL  C56   75  -1.970  -0.438   0.110
+    1  MOL   N8   76  -1.753   0.237  -0.081
+    1  MOL  C57   77  -1.718   0.377  -0.066
+    1  MOL  C58   78  -1.843   0.458  -0.107
+    1  MOL  O13   79  -1.942   0.403  -0.149
+    1  MOL  O14   80  -1.835   0.590  -0.094
+    1  MOL  C59   81  -1.590   0.411  -0.147
+    1  MOL  C60   82  -1.521   0.538  -0.103
+    1  MOL  O15   83  -1.575   0.629  -0.043
+    1  MOL  O16   84  -1.391   0.550  -0.134
+    1  MOL   H1   85   2.204  -0.492  -0.226
+    1  MOL   H2   86   1.906  -0.266  -0.218
+    1  MOL   H3   87   2.010  -0.159  -0.312
+    1  MOL   H4   88   2.270  -0.272  -0.125
+    1  MOL   H5   89   1.992   0.026  -0.056
+    1  MOL   H6   90   1.595  -0.123   0.052
+    1  MOL   H7   91   1.783  -0.050   0.254
+    1  MOL   H8   92   1.903   0.066   0.197
+    1  MOL   H9   93   1.580   0.025   0.364
+    1  MOL  H10   94   1.839   0.312   0.177
+    1  MOL  H11   95   1.452   0.209   0.469
+    1  MOL  H12   96   1.712   0.496   0.281
+    1  MOL  H13   97   1.517   0.445   0.428
+    1  MOL  H14   98   1.454  -0.196  -0.128
+    1  MOL  H15   99   1.416  -0.068  -0.245
+    1  MOL  H16  100   1.285  -0.156  -0.164
+    1  MOL  H17  101   1.331   0.187  -0.087
+    1  MOL  H18  102   1.085   0.241  -0.183
+    1  MOL  H19  103   1.110   0.294  -0.017
+    1  MOL  H20  104   0.916   0.086  -0.022
+    1  MOL  H21  105  -1.053  -0.179  -0.020
+    1  MOL  H22  106  -1.267  -0.217   0.164
+    1  MOL  H23  107  -1.213  -0.373   0.103
+    1  MOL  H24  108  -1.454  -0.321   0.028
+    1  MOL  H25  109  -1.440  -0.468  -0.243
+    1  MOL  H26  110  -1.587  -0.468  -0.143
+    1  MOL  H27  111  -1.594  -0.409  -0.310
+    1  MOL  H28  112  -1.544  -0.080  -0.006
+    1  MOL  H29  113  -1.901   0.040   0.096
+    1  MOL  H30  114  -1.754  -0.013   0.177
+    1  MOL  H31  115  -1.844   0.221  -0.123
+    1  MOL  H32  116  -1.745   0.617  -0.061
+    1  MOL  H33  117  -1.613   0.419  -0.254
+    1  MOL  H34  118  -1.522   0.327  -0.134
+    1  MOL  H35  119  -1.358   0.469  -0.176
+    1  MOL  H36  120   2.057  -0.249  -0.021
+    1  MOL  H37  121   1.766   0.105  -0.010
+    1  MOL  H38  122   1.350  -0.059   0.055
+    1  MOL  H39  123   0.736   0.071   0.085
+    1  MOL  H40  124   0.503   0.007   0.117
+    1  MOL  H41  125   0.381   0.299  -0.175
+    1  MOL  H42  126   0.619   0.364  -0.207
+    1  MOL  H43  127   0.214   0.174  -0.073
+    1  MOL  H44  128   0.296  -0.045   0.125
+    1  MOL  H45  129   0.005   0.133  -0.053
+    1  MOL  H46  130  -0.227   0.069  -0.022
+    1  MOL  H47  131  -0.106  -0.244   0.249
+    1  MOL  H48  132   0.129  -0.178   0.218
+    1  MOL  H49  133  -0.333  -0.220   0.208
+    1  MOL  H50  134  -0.426  -0.003   0.012
+    1  MOL  H51  135  -0.532  -0.285   0.251
+    1  MOL  H52  136  -0.769  -0.350   0.280
+    1  MOL  H53  137  -0.652   0.015  -0.033
+    1  MOL  H54  138  -0.887  -0.049  -0.003
+    1  MOL  H55  139  -1.436  -0.222  -0.249
+    1  MOL  H56  140  -1.803  -0.014  -0.126
+    1  MOL  H57  141  -2.031  -0.141  -0.045
+    1  MOL  H58  142  -2.116  -0.373  -0.034
+    1  MOL  H59  143  -2.008  -0.540   0.115
+    1  MOL  H60  144  -1.813  -0.472   0.254
+    1  MOL  H61  145  -1.728  -0.240   0.244
+    1  MOL  H62  146  -1.696   0.398   0.040
+    4.41340     4.16860     4.86170
